rename webclient to components
